test/bbackupd: add specialised versions of some tests for S3
Not all tests are specialised because many are quite slow, and would behave in
the same way, not expected to exercise any additional S3-related code paths
that are not covered by the tests that have been implemented.

+++ lib/common/Test.h
@@ -268,7 +268,7 @@ void sync_and_wait();
 void terminate_bbackupd(int pid);
 
 // Wait a given number of seconds for something to complete
-void wait_for_operation(int seconds, const char* message);
+void wait_for_operation(float seconds, const char* message);
 void safe_sleep(int seconds);
 std::auto_ptr<Configuration> load_config_file(const std::string& config_file,
 	const ConfigurationVerify& verify);
